refactor: move user.. files into own pkg -> user
