+++ pkg/user/userHandler_test.go
@@ -1,4 +1,4 @@
-package main
+package user
 
 import (
 	"net/http"
